// src/app/share/models/players.ts
export interface Players {
  PlayerID: number;
  TeamID: number;
  Team: string;
  FirstName: string;
  LastName: string;
  Position: string;
  BirthDate: string;
  Height: string;
  Weight: string;
  BirthCity: string;
  BirthCountry: string;
  College: string;
  Salary: number;
  Experience: number;
  StatsPlayerID: number;
  NbaDotComPlayerID: number;
}